refactor(mcp): extract config parsing functions to moltis-mcp
Move parse_server_config, parse_string_map, parse_secret_string_map,
and merge_env_overrides from gateway's mcp_service.rs to a new
config_parsing module in the moltis-mcp crate.

These are pure functions with no gateway state dependency — they only
operate on serde_json::Value and moltis-mcp types.

Changes:
- Add crates/mcp/src/config_parsing.rs with functions + 12 tests
- Re-export parse_server_config and merge_env_overrides from moltis-mcp
- Gateway mcp_service.rs re-exports via pub(crate) use
- Remove duplicate merge_env_overrides from server.rs
- Map moltis_mcp::Error → ServiceError at gateway call sites

Net reduction: ~200 lines from gateway.
