+++ crates/gateway/src/mcp_service.rs
@@ -5,7 +5,6 @@ use std::{collections::HashMap, sync::Arc};
 use {
     anyhow::Result,
     async_trait::async_trait,
-    secrecy::{ExposeSecret, Secret},
     serde_json::Value,
     tokio::sync::RwLock,
     tracing::{info, warn},
@@ -17,6 +16,9 @@ use {
 };
 
 use crate::services::{McpService, ServiceError, ServiceResult};
+
+// Re-export pure parsing functions that now live in moltis-mcp.
+pub(crate) use moltis_mcp::{merge_env_overrides, parse_server_config};
 
 // ── McpToolAdapter: bridge McpAgentTool → AgentTool ─────────────────────────
 
@@ -76,189 +78,6 @@ pub async fn sync_mcp_tools(
 // ── Config parsing helper ───────────────────────────────────────────────────
 
 /// Extract an `McpServerConfig` from JSON params.
-///
-/// For updates, omitted fields inherit from `existing`.
-fn parse_server_config(
-    params: &Value,
-    existing: Option<&moltis_mcp::McpServerConfig>,
-) -> Result<moltis_mcp::McpServerConfig, ServiceError> {
-    let transport = match params.get("transport").and_then(|v| v.as_str()) {
-        Some("sse") => moltis_mcp::TransportType::Sse,
-        Some("streamable-http" | "streamable_http" | "http") => {
-            moltis_mcp::TransportType::StreamableHttp
-        },
-        Some(_) => moltis_mcp::TransportType::Stdio,
-        None => existing
-            .map(|cfg| cfg.transport)
-            .unwrap_or(moltis_mcp::TransportType::Stdio),
-    };
-
-    let command = params
-        .get("command")
-        .and_then(|v| v.as_str())
-        .map(String::from)
-        .or_else(|| existing.map(|cfg| cfg.command.clone()))
-        .unwrap_or_default();
-
-    if matches!(transport, moltis_mcp::TransportType::Stdio) && command.trim().is_empty() {
-        return Err(ServiceError::message("missing 'command' parameter"));
-    }
-
-    let args: Vec<String> = if params.get("args").is_some() {
-        params
-            .get("args")
-            .and_then(|v| serde_json::from_value(v.clone()).ok())
-            .unwrap_or_default()
-    } else {
-        existing.map(|cfg| cfg.args.clone()).unwrap_or_default()
-    };
-
-    let enabled = params
-        .get("enabled")
-        .and_then(|v| v.as_bool())
-        .or_else(|| existing.map(|cfg| cfg.enabled))
-        .unwrap_or(true);
-
-    let request_timeout_secs = if let Some(v) = params.get("request_timeout_secs") {
-        if v.is_null() {
-            None
-        } else {
-            let secs = v
-                .as_u64()
-                .ok_or_else(|| ServiceError::message("invalid 'request_timeout_secs' parameter"))?;
-            if secs == 0 {
-                return Err(ServiceError::message(
-                    "'request_timeout_secs' must be greater than 0",
-                ));
-            }
-            Some(secs)
-        }
-    } else {
-        existing.and_then(|cfg| cfg.request_timeout_secs)
-    };
-
-    let url = if params.get("url").is_some() {
-        if params.get("url").is_some_and(Value::is_null) {
-            None
-        } else {
-            params
-                .get("url")
-                .and_then(|v| v.as_str())
-                .map(|value| Secret::new(value.to_string()))
-        }
-    } else {
-        existing.and_then(|cfg| cfg.url.clone())
-    };
-
-    let headers = if matches!(
-        transport,
-        moltis_mcp::TransportType::Sse | moltis_mcp::TransportType::StreamableHttp
-    ) {
-        if params.get("headers").is_some() {
-            parse_secret_string_map(params.get("headers").unwrap_or(&Value::Null))
-        } else {
-            existing.map(|cfg| cfg.headers.clone()).unwrap_or_default()
-        }
-    } else {
-        HashMap::new()
-    };
-
-    let env = if matches!(
-        transport,
-        moltis_mcp::TransportType::Sse | moltis_mcp::TransportType::StreamableHttp
-    ) {
-        HashMap::new()
-    } else if params.get("env").is_some() {
-        parse_string_map(params.get("env").unwrap_or(&Value::Null))
-    } else {
-        existing.map(|cfg| cfg.env.clone()).unwrap_or_default()
-    };
-
-    if matches!(
-        transport,
-        moltis_mcp::TransportType::Sse | moltis_mcp::TransportType::StreamableHttp
-    ) && url
-        .as_ref()
-        .map(ExposeSecret::expose_secret)
-        .is_none_or(|candidate| candidate.trim().is_empty())
-    {
-        return Err(ServiceError::message(format!(
-            "missing 'url' parameter for '{}' transport",
-            transport
-        )));
-    }
-
-    let oauth = if let Some(v) = params.get("oauth") {
-        if v.is_null() {
-            None
-        } else {
-            let client_id = v
-                .get("client_id")
-                .and_then(|val| val.as_str())
-                .ok_or_else(|| ServiceError::message("missing 'oauth.client_id' parameter"))?
-                .to_string();
-            let auth_url = v
-                .get("auth_url")
-                .and_then(|val| val.as_str())
-                .ok_or_else(|| ServiceError::message("missing 'oauth.auth_url' parameter"))?
-                .to_string();
-            let token_url = v
-                .get("token_url")
-                .and_then(|val| val.as_str())
-                .ok_or_else(|| ServiceError::message("missing 'oauth.token_url' parameter"))?
-                .to_string();
-            let scopes: Vec<String> = v
-                .get("scopes")
-                .and_then(|s| serde_json::from_value(s.clone()).ok())
-                .unwrap_or_default();
-            Some(moltis_mcp::registry::McpOAuthConfig {
-                client_id,
-                auth_url,
-                token_url,
-                scopes,
-            })
-        }
-    } else {
-        existing.and_then(|cfg| cfg.oauth.clone())
-    };
-
-    let display_name = match params.get("display_name") {
-        Some(v) if v.is_null() => None,
-        Some(v) => v.as_str().map(String::from),
-        None => existing.and_then(|cfg| cfg.display_name.clone()),
-    };
-
-    Ok(moltis_mcp::McpServerConfig {
-        command,
-        args,
-        env,
-        enabled,
-        request_timeout_secs,
-        transport,
-        url: if matches!(
-            transport,
-            moltis_mcp::TransportType::Sse | moltis_mcp::TransportType::StreamableHttp
-        ) {
-            url
-        } else {
-            None
-        },
-        headers,
-        oauth,
-        display_name,
-    })
-}
-
-fn parse_string_map(value: &Value) -> HashMap<String, String> {
-    serde_json::from_value(value.clone()).unwrap_or_default()
-}
-
-fn parse_secret_string_map(value: &Value) -> HashMap<String, Secret<String>> {
-    parse_string_map(value)
-        .into_iter()
-        .map(|(key, value)| (key, Secret::new(value)))
-        .collect()
-}
 
 // ── LiveMcpService ──────────────────────────────────────────────────────────
 
@@ -327,21 +146,6 @@ impl LiveMcpService {
     }
 }
 
-fn merge_env_overrides(
-    base_overrides: &HashMap<String, String>,
-    additional: Vec<(String, String)>,
-) -> HashMap<String, String> {
-    // Config `[env]` values stay authoritative so checked-in config cannot be
-    // silently shadowed by mutable UI-managed entries from the credential store.
-    let mut merged = base_overrides.clone();
-    for (key, value) in additional {
-        if key.trim().is_empty() || value.trim().is_empty() {
-            continue;
-        }
-        merged.entry(key).or_insert(value);
-    }
-    merged
-}
 
 #[async_trait]
 impl McpService for LiveMcpService {
@@ -361,7 +165,8 @@ impl McpService for LiveMcpService {
             .map(str::trim)
             .filter(|v| !v.is_empty())
             .map(ToOwned::to_owned);
-        let config = parse_server_config(&params, None)?;
+        let config = parse_server_config(&params, None)
+            .map_err(|e| ServiceError::message(e.to_string()))?;
         self.refresh_manager_env_overrides().await;
 
         // If a server with this name already exists, append a numeric suffix.
@@ -552,7 +357,8 @@ impl McpService for LiveMcpService {
             .get(name)
             .cloned()
             .ok_or_else(|| format!("MCP server '{name}' not found"))?;
-        let config = parse_server_config(&params, Some(&existing))?;
+        let config = parse_server_config(&params, Some(&existing))
+            .map_err(|e| ServiceError::message(e.to_string()))?;
         self.refresh_manager_env_overrides().await;
 
         self.manager
@@ -656,331 +462,6 @@ mod tests {
         secrecy::{ExposeSecret, Secret},
     };
 
-    #[test]
-    fn parse_server_config_allows_sse_without_command() {
-        let cfg = parse_server_config(
-            &serde_json::json!({
-                "transport": "sse",
-                "url": "https://mcp.linear.app/mcp",
-                "enabled": true
-            }),
-            None,
-        );
-        assert!(
-            cfg.is_ok(),
-            "expected SSE config to parse without command, got: {cfg:?}"
-        );
-        let Ok(cfg) = cfg else {
-            panic!("SSE config unexpectedly failed to parse");
-        };
-
-        assert!(matches!(cfg.transport, moltis_mcp::TransportType::Sse));
-        assert_eq!(cfg.command, "");
-        assert_eq!(
-            cfg.url
-                .as_ref()
-                .map(ExposeSecret::expose_secret)
-                .map(String::as_str),
-            Some("https://mcp.linear.app/mcp")
-        );
-    }
-
-    #[test]
-    fn parse_server_config_requires_command_for_stdio() {
-        let err = parse_server_config(
-            &serde_json::json!({
-                "transport": "stdio",
-                "args": ["-y", "@modelcontextprotocol/server-filesystem"]
-            }),
-            None,
-        )
-        .err();
-
-        assert_eq!(
-            err.as_ref().map(ToString::to_string).as_deref(),
-            Some("missing 'command' parameter")
-        );
-    }
-
-    #[test]
-    fn parse_server_config_requires_url_for_sse() {
-        let err = parse_server_config(
-            &serde_json::json!({
-                "transport": "sse",
-            }),
-            None,
-        )
-        .err();
-
-        assert_eq!(
-            err.as_ref().map(ToString::to_string).as_deref(),
-            Some("missing 'url' parameter for 'sse' transport")
-        );
-    }
-
-    #[test]
-    fn parse_server_config_allows_streamable_http_without_command() {
-        let cfg = parse_server_config(
-            &serde_json::json!({
-                "transport": "streamable-http",
-                "url": "https://mcp.example.com/mcp",
-                "enabled": true
-            }),
-            None,
-        );
-        assert!(
-            cfg.is_ok(),
-            "expected Streamable HTTP config to parse without command, got: {cfg:?}"
-        );
-        let Ok(cfg) = cfg else {
-            panic!("Streamable HTTP config unexpectedly failed to parse");
-        };
-
-        assert!(matches!(
-            cfg.transport,
-            moltis_mcp::TransportType::StreamableHttp
-        ));
-        assert_eq!(cfg.command, "");
-        assert_eq!(
-            cfg.url
-                .as_ref()
-                .map(ExposeSecret::expose_secret)
-                .map(String::as_str),
-            Some("https://mcp.example.com/mcp")
-        );
-    }
-
-    #[test]
-    fn parse_server_config_requires_url_for_streamable_http() {
-        let err = parse_server_config(
-            &serde_json::json!({
-                "transport": "streamable-http",
-            }),
-            None,
-        )
-        .err();
-
-        assert_eq!(
-            err.as_ref().map(ToString::to_string).as_deref(),
-            Some("missing 'url' parameter for 'streamable-http' transport")
-        );
-    }
-
-    #[test]
-    fn parse_server_config_update_preserves_existing_sse_fields() {
-        let existing = moltis_mcp::McpServerConfig {
-            transport: moltis_mcp::TransportType::Sse,
-            url: Some(Secret::new("https://mcp.linear.app/mcp".to_string())),
-            ..Default::default()
-        };
-
-        let cfg = parse_server_config(
-            &serde_json::json!({
-                "enabled": false
-            }),
-            Some(&existing),
-        );
-        assert!(
-            cfg.is_ok(),
-            "expected parser to preserve SSE defaults from existing config, got: {cfg:?}"
-        );
-        let Ok(cfg) = cfg else {
-            panic!("failed to parse update with inherited SSE config");
-        };
-
-        assert!(matches!(cfg.transport, moltis_mcp::TransportType::Sse));
-        assert_eq!(
-            cfg.url
-                .as_ref()
-                .map(ExposeSecret::expose_secret)
-                .map(String::as_str),
-            Some("https://mcp.linear.app/mcp")
-        );
-        assert!(!cfg.enabled);
-    }
-
-    #[test]
-    fn parse_server_config_update_preserves_oauth_when_omitted() {
-        let existing = moltis_mcp::McpServerConfig {
-            transport: moltis_mcp::TransportType::Sse,
-            url: Some(Secret::new("https://mcp.linear.app/mcp".to_string())),
-            oauth: Some(moltis_mcp::McpOAuthConfig {
-                client_id: "linear-client".to_string(),
-                auth_url: "https://linear.app/oauth/authorize".to_string(),
-                token_url: "https://api.linear.app/oauth/token".to_string(),
-                scopes: vec!["read".to_string(), "write".to_string()],
-            }),
-            ..Default::default()
-        };
-
-        let cfg = parse_server_config(
-            &serde_json::json!({
-                "transport": "sse"
-            }),
-            Some(&existing),
-        );
-        assert!(
-            cfg.is_ok(),
-            "expected parser to preserve existing oauth fields, got: {cfg:?}"
-        );
-        let Ok(cfg) = cfg else {
-            panic!("failed to parse update while preserving oauth");
-        };
-
-        assert!(cfg.oauth.is_some(), "expected oauth to be preserved");
-        let Some(oauth) = cfg.oauth else {
-            panic!("oauth missing after parse");
-        };
-        assert_eq!(oauth.client_id, "linear-client");
-        assert_eq!(oauth.auth_url, "https://linear.app/oauth/authorize");
-        assert_eq!(oauth.token_url, "https://api.linear.app/oauth/token");
-        assert_eq!(oauth.scopes, vec!["read".to_string(), "write".to_string()]);
-    }
-
-    #[test]
-    fn parse_server_config_preserves_and_replaces_sse_headers() {
-        let existing = moltis_mcp::McpServerConfig {
-            transport: moltis_mcp::TransportType::Sse,
-            url: Some(Secret::new("https://mcp.linear.app/mcp".to_string())),
-            headers: HashMap::from([(
-                "Authorization".to_string(),
-                Secret::new("Bearer old-secret".to_string()),
-            )]),
-            ..Default::default()
-        };
-
-        let preserved = parse_server_config(
-            &serde_json::json!({
-                "transport": "sse"
-            }),
-            Some(&existing),
-        );
-        assert!(
-            preserved.is_ok(),
-            "expected header preservation, got: {preserved:?}"
-        );
-        let Ok(preserved) = preserved else {
-            panic!("header preservation unexpectedly failed");
-        };
-        assert_eq!(preserved.headers.len(), 1);
-        assert_eq!(
-            preserved
-                .headers
-                .get("Authorization")
-                .map(ExposeSecret::expose_secret)
-                .map(String::as_str),
-            Some("Bearer old-secret")
-        );
-
-        let replaced = parse_server_config(
-            &serde_json::json!({
-                "transport": "sse",
-                "headers": {
-                    "X-Workspace": "team-alpha"
-                }
-            }),
-            Some(&existing),
-        );
-        assert!(
-            replaced.is_ok(),
-            "expected header replacement, got: {replaced:?}"
-        );
-        let Ok(replaced) = replaced else {
-            panic!("header replacement unexpectedly failed");
-        };
-        assert_eq!(replaced.headers.len(), 1);
-        assert_eq!(
-            replaced
-                .headers
-                .get("X-Workspace")
-                .map(ExposeSecret::expose_secret)
-                .map(String::as_str),
-            Some("team-alpha")
-        );
-    }
-
-    #[test]
-    fn parse_server_config_allows_clearing_sse_headers() {
-        let existing = moltis_mcp::McpServerConfig {
-            transport: moltis_mcp::TransportType::Sse,
-            url: Some(Secret::new("https://mcp.linear.app/mcp".to_string())),
-            headers: HashMap::from([(
-                "Authorization".to_string(),
-                Secret::new("Bearer old-secret".to_string()),
-            )]),
-            ..Default::default()
-        };
-
-        let cleared = parse_server_config(
-            &serde_json::json!({
-                "transport": "sse",
-                "headers": {}
-            }),
-            Some(&existing),
-        );
-        assert!(
-            cleared.is_ok(),
-            "expected header clearing, got: {cleared:?}"
-        );
-        let Ok(cleared) = cleared else {
-            panic!("header clearing unexpectedly failed");
-        };
-        assert!(cleared.headers.is_empty());
-    }
-
-    #[test]
-    fn merge_env_overrides_keeps_config_values_authoritative() {
-        let base = HashMap::from([
-            ("OPENAI_API_KEY".to_string(), "config-openai".to_string()),
-            ("BRAVE_API_KEY".to_string(), "config-brave".to_string()),
-        ]);
-
-        let merged = merge_env_overrides(&base, vec![
-            ("OPENAI_API_KEY".to_string(), "ui-openai".to_string()),
-            (
-                "PERPLEXITY_API_KEY".to_string(),
-                "ui-perplexity".to_string(),
-            ),
-        ]);
-
-        assert_eq!(
-            merged.get("OPENAI_API_KEY").map(String::as_str),
-            Some("config-openai")
-        );
-        assert_eq!(
-            merged.get("PERPLEXITY_API_KEY").map(String::as_str),
-            Some("ui-perplexity")
-        );
-        assert_eq!(
-            merged.get("BRAVE_API_KEY").map(String::as_str),
-            Some("config-brave")
-        );
-    }
-
-    #[test]
-    #[allow(clippy::unwrap_used, clippy::expect_used)]
-    fn parse_server_config_preserves_request_timeout_override() {
-        let existing = moltis_mcp::McpServerConfig {
-            command: "npx".to_string(),
-            args: vec![
-                "-y".to_string(),
-                "@modelcontextprotocol/server-memory".to_string(),
-            ],
-            request_timeout_secs: Some(90),
-            ..Default::default()
-        };
-
-        let cfg = parse_server_config(
-            &serde_json::json!({
-                "enabled": false
-            }),
-            Some(&existing),
-        )
-        .expect("expected request timeout override to be preserved");
-
-        assert_eq!(cfg.request_timeout_secs, Some(90));
-        assert!(!cfg.enabled);
-    }
 
     #[tokio::test]
     async fn test_sync_mcp_tools_empty_manager() {
